Udacity-iOS-Nanodegree/project/TripJournal/TripJournalAPI/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.5.2
sqlalchemy==1.4.50
python-jose[cryptography]==3.3.0
databases==0.8.0
bcrypt==4.0.1
passlib==1.7.4
python-multipart==0.0.6
slowapi==0.1.8
pillow==10.1.0